Refactor chat with entry screen for reuse
